refactor: move standard VM to own module
I'm planning to add optimized virtual machine implementation
later. The optimized implementation will be able to jump to the
specified memory address instead of searching loop bounds each time.
It will also increment and decrement values with arbitrary numbers.

To support this there is a trait for virtual machine in place.